requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.30.1
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
schedule==1.2.2
sniffio==1.3.1
soupsieve==2.5
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
